# Optimize bitwriter code lookups