--- a/flate/huffman_bit_writer.go
+++ b/flate/huffman_bit_writer.go
@@ -286,11 +286,10 @@ func (w *huffmanBitWriter) writeCode(code *huffmanEncoder, literal uint32) {
 }
 */
 
-func (w *huffmanBitWriter) writeCode(code *huffmanEncoder, literal uint32) {
+func (w *huffmanBitWriter) writeCode(c hcode) {
 	if w.err != nil {
 		return
 	}
-	c := code.codes[literal]
 	w.bits |= uint64(c.code()) << w.nbits
 	w.nbits += c.bits()
 	if w.nbits >= 48 {
@@ -345,7 +344,7 @@ func (w *huffmanBitWriter) writeDynamicHeader(numLiterals int, numOffsets int, n
 			break
 		}
 		// The low byte contains the actual code to generate.
-		w.writeCode(w.codegenEncoding, uint32(codeWord))
+		w.writeCode(w.codegenEncoding.codes[uint32(codeWord)])
 
 		switch codeWord {
 		case 16:
@@ -406,10 +405,9 @@ func (w *huffmanBitWriter) writeBlock(tok tokens, eof bool, input []byte) {
 	tokens := tok.tokens[0 : tok.n+1]
 
 	for _, t := range tokens {
-		switch t.typ() {
-		case literalType:
+		if t < matchType {
 			w.literalFreq[t.literal()]++
-		case matchType:
+		} else {
 			length := t.length()
 			offset := t.offset()
 			w.literalFreq[lengthCodesStart+lengthCode(length)]++
@@ -508,16 +506,17 @@ func (w *huffmanBitWriter) writeBlock(tok tokens, eof bool, input []byte) {
 	} else {
 		w.writeDynamicHeader(numLiterals, numOffsets, numCodegens, eof)
 	}
+
+	leCodes := literalEncoding.codes
+	oeCodes := offsetEncoding.codes
 	for _, t := range tokens {
-		switch t.typ() {
-		case literalType:
-			w.writeCode(literalEncoding, t.literal())
-			break
-		case matchType:
+		if t < matchType {
+			w.writeCode(leCodes[t.literal()])
+		} else {
 			// Write the length
 			length := t.length()
 			lengthCode := lengthCode(length)
-			w.writeCode(literalEncoding, lengthCode+lengthCodesStart)
+			w.writeCode(leCodes[lengthCode+lengthCodesStart])
 			extraLengthBits := uint(lengthExtraBits[lengthCode])
 			if extraLengthBits > 0 {
 				extraLength := int32(length - lengthBase[lengthCode])
@@ -526,15 +525,12 @@ func (w *huffmanBitWriter) writeBlock(tok tokens, eof bool, input []byte) {
 			// Write the offset
 			offset := t.offset()
 			offsetCode := offsetCode(offset)
-			w.writeCode(offsetEncoding, offsetCode)
+			w.writeCode(oeCodes[offsetCode])
 			extraOffsetBits := uint(offsetExtraBits[offsetCode])
 			if extraOffsetBits > 0 {
 				extraOffset := int32(offset - offsetBase[offsetCode])
 				w.writeBits(extraOffset, extraOffsetBits)
 			}
-			break
-		default:
-			panic("unknown token type: " + string(t))
 		}
 	}
 }
@@ -557,10 +553,9 @@ func (w *huffmanBitWriter) writeBlockDynamic(tok tokens, eof bool, input []byte)
 	tokens := tok.tokens[0 : tok.n+1]
 
 	for _, t := range tokens {
-		switch t.typ() {
-		case literalType:
+		if t < matchType {
 			w.literalFreq[t.literal()]++
-		case matchType:
+		} else {
 			length := t.length()
 			offset := t.offset()
 			w.literalFreq[lengthCodesStart+lengthCode(length)]++
@@ -603,16 +598,17 @@ func (w *huffmanBitWriter) writeBlockDynamic(tok tokens, eof bool, input []byte)
 
 	// Write Huffman table.
 	w.writeDynamicHeader(numLiterals, numOffsets, numCodegens, eof)
+	leCodes := literalEncoding.codes
+	oeCodes := offsetEncoding.codes
+
 	for _, t := range tokens {
-		switch t.typ() {
-		case literalType:
-			w.writeCode(literalEncoding, t.literal())
-			break
-		case matchType:
+		if t < matchType {
+			w.writeCode(leCodes[t.literal()])
+		} else {
 			// Write the length
 			length := t.length()
 			lengthCode := lengthCode(length)
-			w.writeCode(literalEncoding, lengthCode+lengthCodesStart)
+			w.writeCode(leCodes[lengthCode+lengthCodesStart])
 			extraLengthBits := uint(lengthExtraBits[lengthCode])
 			if extraLengthBits > 0 {
 				extraLength := int32(length - lengthBase[lengthCode])
@@ -621,15 +617,12 @@ func (w *huffmanBitWriter) writeBlockDynamic(tok tokens, eof bool, input []byte)
 			// Write the offset
 			offset := t.offset()
 			offsetCode := offsetCode(offset)
-			w.writeCode(offsetEncoding, offsetCode)
+			w.writeCode(oeCodes[offsetCode])
 			extraOffsetBits := uint(offsetExtraBits[offsetCode])
 			if extraOffsetBits > 0 {
 				extraOffset := int32(offset - offsetBase[offsetCode])
 				w.writeBits(extraOffset, extraOffsetBits)
 			}
-			break
-		default:
-			panic("unknown token type: " + string(t))
 		}
 	}
 }
@@ -704,9 +697,10 @@ func (w *huffmanBitWriter) writeBlockHuff(eof bool, input []byte) {
 
 	// Huffman.
 	w.writeDynamicHeader(numLiterals, numOffsets, numCodegens, eof)
+	encoding := w.literalEncoding.codes
 	for _, t := range input {
 		// Bitwriting inlined, ~30% speedup
-		c := w.literalEncoding.codes[t]
+		c := encoding[t]
 		w.bits |= uint64(c.code()) << w.nbits
 		w.nbits += c.bits()
 		if w.nbits >= 48 {
@@ -732,5 +726,5 @@ func (w *huffmanBitWriter) writeBlockHuff(eof bool, input []byte) {
 			}
 		}
 	}
-	w.writeCode(w.literalEncoding, endBlockMarker)
+	w.writeCode(encoding[endBlockMarker])
 }