Optimize bitwriter, by eliminating some nil checks and use a simpler test to determine the token type.

```
name                       old time/op    new time/op    delta
EncodeDigitsConstant1e4-4   159MB/s ± 0%   184MB/s ± 2%  +15.30%
EncodeDigitsConstant1e5-4   173MB/s ± 0%   186MB/s ± 7%   +7.21%
EncodeDigitsConstant1e6-4   171MB/s ± 0%   195MB/s ± 1%  +13.94%
EncodeDigitsSpeed1e4-4     45.4MB/s ± 1%  47.5MB/s ± 2%   +4.60%
EncodeDigitsSpeed1e5-4     70.5MB/s ± 0%  71.7MB/s ± 2%   +1.72%
EncodeDigitsSpeed1e6-4      110MB/s ± 0%   118MB/s ± 0%   +7.96%
EncodeDigitsDefault1e4-4   26.6MB/s ± 0%  26.3MB/s ± 0%   -0.83%
EncodeDigitsDefault1e5-4   21.3MB/s ± 1%  22.0MB/s ± 0%   +3.00%
EncodeDigitsDefault1e6-4   19.7MB/s ± 0%  21.2MB/s ± 0%   +8.03%
EncodeDigitsCompress1e4-4  22.9MB/s ± 1%  22.8MB/s ± 2%   -0.65%
EncodeDigitsCompress1e5-4  16.7MB/s ± 0%  16.9MB/s ± 1%   +1.14%
EncodeDigitsCompress1e6-4  15.3MB/s ± 0%  16.2MB/s ± 1%   +5.75%
EncodeTwainConstant1e4-4    117MB/s ± 1%   128MB/s ± 1%   +9.73%
EncodeTwainConstant1e5-4    149MB/s ± 0%   169MB/s ± 1%  +13.13%
EncodeTwainConstant1e6-4    146MB/s ± 2%   167MB/s ± 1%  +14.26%
EncodeTwainSpeed1e4-4      43.0MB/s ± 0%  44.0MB/s ± 1%   +2.54%
EncodeTwainSpeed1e5-4      50.5MB/s ± 0%  52.4MB/s ± 1%   +3.61%
EncodeTwainSpeed1e6-4      52.0MB/s ± 0%  53.9MB/s ± 0%   +3.71%
EncodeTwainDefault1e4-4    25.2MB/s ± 2%  25.6MB/s ± 2%   +1.47%
EncodeTwainDefault1e5-4    28.0MB/s ± 1%  27.7MB/s ± 0%   -1.07%
EncodeTwainDefault1e6-4    28.0MB/s ± 0%  27.2MB/s ± 5%   -2.86%
EncodeTwainCompress1e4-4   20.2MB/s ± 0%  20.1MB/s ± 2%   -0.64%
EncodeTwainCompress1e5-4   13.5MB/s ± 0%  13.5MB/s ± 0%   -0.48%
EncodeTwainCompress1e6-4   12.4MB/s ± 1%  12.0MB/s ± 3%   -2.87%
```